backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
python-multipart==0.0.6
pydantic==2.5.1
python-dotenv==1.0.0
pillow==11.0.0
stripe==12.5.1
PyJWT==2.9.0
requests==2.32.3
supabase==2.9.0
